+++ posthog/session_recordings/session_recording_api.py
@@ -5,7 +5,6 @@ from datetime import datetime, timedelta, timezone
 import json
 from typing import Any, List, Type, cast, Dict, Tuple
 
-import openai
 from django.conf import settings
 
 import posthoganalytics
@@ -25,8 +24,7 @@ from posthog.api.routing import StructuredViewSetMixin
 from posthog.auth import SharingAccessTokenAuthentication
 from posthog.cloud_utils import is_cloud
 from posthog.constants import SESSION_RECORDINGS_FILTER_IDS
-from posthog.models import User, Team
-from posthog.models.element import chain_to_elements
+from posthog.models import User
 from posthog.models.filters.session_recordings_filter import SessionRecordingsFilter
 from posthog.models.person.person import PersonDistinctId
 from posthog.session_recordings.models.session_recording import SessionRecording
@@ -52,13 +50,13 @@ from posthog.rate_limit import (
 )
 from posthog.session_recordings.queries.session_replay_events import SessionReplayEvents
 from posthog.session_recordings.realtime_snapshots import get_realtime_snapshots, publish_subscription
+from posthog.session_recordings.session_summary.summarize_session import summarize_recording
 from posthog.session_recordings.snapshots.convert_legacy_snapshots import (
     convert_original_version_lts_recording,
 )
 from posthog.storage import object_storage
 from prometheus_client import Counter
 
-from posthog.utils import get_instance_region
 
 SNAPSHOT_SOURCE_REQUESTED = Counter(
     "session_snapshots_requested_counter",
@@ -180,222 +178,6 @@ def list_recordings_response(
         f"{key};dur={round(duration, ndigits=2)}" for key, duration in timings.items()
     )
     return response
-
-
-def is_boring_string(element: str) -> bool:
-    return element in ["a", "div", "span", "p", "h1", "h2", "h3", "h4", "h5", "h6"]
-
-
-def reduce_elements_chain(session_events: Tuple[List | None, List | None]) -> Tuple[List | None, List | None]:
-    columns, results = session_events
-
-    if columns is None or results is None:
-        return columns, results
-
-    # find elements_chain column index
-    elements_chain_index = None
-    for i, column in enumerate(columns):
-        if column == "elements_chain":
-            elements_chain_index = i
-            break
-
-    reduced_results = []
-    for result in results:
-        if elements_chain_index is None:
-            reduced_results.append(result)
-            continue
-
-        elements_chain: str | None = result[elements_chain_index]
-        if not elements_chain:
-            reduced_results.append(result)
-            continue
-
-        # the elements chain has lots of information that we don't need
-        elements = [e for e in chain_to_elements(elements_chain) if e.tag_name in e.USEFUL_ELEMENTS]
-
-        result_list = list(result)
-        result_list[elements_chain_index] = [{"tag": e.tag_name, "text": e.text, "href": e.href} for e in elements]
-        reduced_results.append(tuple(result_list))
-
-    return columns, reduced_results
-
-
-def simplify_window_id(session_events: Tuple[List | None, List | None]) -> Tuple[List | None, List | None]:
-    columns, results = session_events
-
-    if columns is None or results is None:
-        return columns, results
-
-    # find window_id column index
-    window_id_index = None
-    for i, column in enumerate(columns):
-        if column == "$window_id":
-            window_id_index = i
-            break
-
-    window_id_mapping: Dict[str, int] = {}
-    simplified_results = []
-    for result in results:
-        if window_id_index is None:
-            simplified_results.append(result)
-            continue
-
-        window_id: str | None = result[window_id_index]
-        if not window_id:
-            simplified_results.append(result)
-            continue
-
-        if window_id not in window_id_mapping:
-            window_id_mapping[window_id] = len(window_id_mapping) + 1
-
-        result_list = list(result)
-        result_list[window_id_index] = window_id_mapping[window_id]
-        simplified_results.append(tuple(result_list))
-
-    return columns, simplified_results
-
-
-def deduplicate_urls(
-    session_events: Tuple[List | None, List | None]
-) -> Tuple[List | None, List | None, Dict[str, str]]:
-    columns, results = session_events
-
-    if columns is None or results is None:
-        return columns, results, {}
-
-    # find url column index
-    url_index = None
-    for i, column in enumerate(columns):
-        if column == "$current_url":
-            url_index = i
-            break
-
-    url_mapping: Dict[str, str] = {}
-    deduplicated_results = []
-    for result in results:
-        if url_index is None:
-            deduplicated_results.append(result)
-            continue
-
-        url: str | None = result[url_index]
-        if not url:
-            deduplicated_results.append(result)
-            continue
-
-        if url not in url_mapping:
-            url_mapping[url] = f"url_{len(url_mapping) + 1}"
-
-        result_list = list(result)
-        result_list[url_index] = url_mapping[url]
-        deduplicated_results.append(tuple(result_list))
-
-    return columns, deduplicated_results, url_mapping
-
-
-def format_dates(session_events: Tuple[List | None, List | None], start: datetime) -> Tuple[List | None, List | None]:
-    columns, results = session_events
-
-    if columns is None or results is None:
-        return columns, results
-
-    # find timestamp column index
-    timestamp_index = None
-    for i, column in enumerate(columns):
-        if column == "timestamp":
-            timestamp_index = i
-            break
-    del columns[timestamp_index]  # remove timestamp column from columns
-    columns.append("milliseconds_since_start")  # add new column to columns at end
-
-    formatted_results = []
-    for result in results:
-        if timestamp_index is None:
-            formatted_results.append(result)
-            continue
-
-        timestamp: datetime | None = result[timestamp_index]
-        if not timestamp:
-            formatted_results.append(result)
-            continue
-
-        result_list = list(result)
-        # remove list item at timestamp_index
-        del result_list[timestamp_index]
-        # insert milliseconds since reference date
-        result_list.append(int((timestamp - start).total_seconds() * 1000))
-        formatted_results.append(tuple(result_list))
-
-    return columns, formatted_results
-
-
-def summarize_recording(recording: SessionRecording, user: User, team: Team):
-    session_metadata = SessionReplayEvents().get_metadata(session_id=str(recording.session_id), team=team)
-    session_events = SessionReplayEvents().get_events(
-        session_id=str(recording.session_id),
-        team=team,
-        metadata=session_metadata,
-        events_to_ignore=[
-            "$feature_flag_called",
-        ],
-    )
-
-    del session_metadata["distinct_id"]
-    start_time = session_metadata["start_time"]
-    session_metadata["start_time"] = start_time.isoformat()
-    session_metadata["end_time"] = session_metadata["end_time"].isoformat()
-
-    session_events_columns, session_events_results, url_mapping = deduplicate_urls(
-        format_dates(reduce_elements_chain(simplify_window_id(session_events)), start=start_time)
-    )
-
-    instance_region = get_instance_region() or "HOBBY"
-    messages = [
-        {
-            "role": "system",
-            "content": """
-            Session Replay is PostHog's tool to record visits to web sites and apps.
-            We also gather events that occur like mouse clicks and key presses.
-            You write two or three sentence concise and simple summaries of those sessions based on a prompt.
-            You are more likely to mention errors or things that look like business success such as checkout events.
-            You don't help with other knowledge.""",
-        },
-        {
-            "role": "user",
-            "content": f"""the session metadata I have is {session_metadata}.
-            it gives an overview of activity and duration""",
-        },
-        {
-            "role": "user",
-            "content": f"""
-            URLs associated with the events can be found in this mapping {url_mapping}.
-            """,
-        },
-        {
-            "role": "user",
-            "content": f"""the session events I have are {session_events_results}.
-            with columns {session_events_columns}.
-            they give an idea of what happened and when,
-            if present the elements_chain extracted from the html can aid in understanding
-            but should not be directly used in your response""",
-        },
-        {
-            "role": "user",
-            "content": """
-            generate a two or three sentence summary of the session.
-            use as concise and simple language as is possible.
-            assume a reading age of around 12 years old.
-            generate no text other than the summary.""",
-        },
-    ]
-    result = openai.ChatCompletion.create(
-        # model="gpt-4-1106-preview",  # allows 128k tokens
-        model="gpt-4",  # allows 8k tokens
-        temperature=0.7,
-        messages=messages,
-        user=f"{instance_region}/{user.pk}",  # The user ID is for tracking within OpenAI in case of overuse/abuse
-    )
-    content: str = result.get("choices", [{}])[0].get("message", {}).get("content", "")
-    return {"ai_result": result, "content": content, "prompt": messages}
 
 
 class SessionRecordingViewSet(StructuredViewSetMixin, viewsets.GenericViewSet):
